Don't alert on expected Karpenter cloudprovider errors
The Karpenter v1 migration docs[1] say:

Karpenter now waits for the underlying instance to be completely terminated
before deleting a node and orchestrates this by emitting
NodeClaimNotFoundError. With this change we expect to see an increase in the
NodeClaimNotFoundError. Customers can filter out this error by label in order
to get accurate values for karpenter_cloudprovider_errors_total metric. Use
this Prometheus filter expression - ({controller!="node.termination"} or
{controller!="nodeclaim.termination"}) and {error!="NodeClaimNotFoundError"}.

This change removes alerts for these expected errors.

1: https://karpenter.sh/v1.0/upgrading/v1-migration